move media tests from deprecated wrapper to MediaOps
Tests for check_content_length, download_blob_from_blossom, and
blossom_client now live in session/groups/media.rs where they call
MediaOps:: directly. Removed all #[cfg(test)]-gated code from the
deprecated groups/media.rs wrapper.
